+++ CHANGELOG.md
@@ -5,7 +5,27 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.16.0](https://github.com/Altinity/ice/compare/v0.16.0...v0.15.0)
 
+### Added
+
+- `ice-rest-catalog` - Add logs to display catalog backend (etcd/sqlite) and timeout for etcd ([#167](https://github.com/Altinity/ice/pull/167))
+- `ice` - Add list-snapshots command ([#163](https://github.com/Altinity/ice/pull/163))
+- `ice` - Add logic to export and import etcd catalog keys as JSON ([#174](https://github.com/Altinity/ice/pull/174))
+
+### Other
+
+- `ice` - Make files command printing logic reusable ([#165](https://github.com/Altinity/ice/pull/165))
+- `ice-rest-catalog` - Added docs on scaling from 1 node to 3. ([#169](https://github.com/Altinity/ice/pull/169))
+`
+## [0.15.0](https://github.com/Altinity/ice/compare/v0.15.0...v0.14.0)
+
+### Added
+
+- `ice`, `ice-rest-catalog` - Added logic to externalize commit retries in rest-catalog. Added metrics for commit retries
+  in rest-catalog. Added logic in insert to handle retries by fetching newer metadata. 
+  Added logic to add Commit Lock using etcd distributed locks
+- 
 ## [0.14.0](https://github.com/Altinity/ice/compare/v0.14.0...v0.13.0)
 
 ### Added